0.5.1: improper charge gain fix (#27)

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Castlebreaker.java
@@ -123,7 +123,7 @@ public class Castlebreaker extends MeleeWeapon {
 	@Override
 	public String status() {
 		if (chargeCap == 100)
-			return Messages.format("%d%%", charge);
+			return Messages.format("%d%%", (int)charge);
 
 
 		//otherwise, if there's no charge, return null.
