+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Echeveria.java
@@ -80,7 +80,7 @@ public class Echeveria extends MeleeWeapon{
     }
 
     @Override
-    public void SPCharge(int value) {
+    public void SPCharge(float value) {
         if (Dungeon.bossLevel()) value = 2;
         super.SPCharge(value);
     }
@@ -191,7 +191,7 @@ public class Echeveria extends MeleeWeapon{
     @Override
     public String status() {
         if (chargeCap == 100)
-            return Messages.format("%d%%", charge);
+            return Messages.format("%d%%", (int)charge);
         return null;
     }
