+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Enfild2.java
@@ -61,7 +61,7 @@ public class Enfild2 extends MeleeWeapon {
         }
         //display as percent
         if (chargeCap == 100)
-            return Messages.format("%d%%", charge);
+            return Messages.format("%d%%", (int)charge);
 
 
         //otherwise, if there's no charge, return null.
