+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Flag.java
@@ -97,7 +97,7 @@ public class Flag extends MeleeWeapon  {
         }
         //display as percent
         if (chargeCap == 100)
-            return Messages.format("%d%%", charge);
+            return Messages.format("%d%%", (int)charge);
 
 
         //otherwise, if there's no charge, return null.
