+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Gluttony.java
@@ -89,7 +89,7 @@ public class Gluttony extends MeleeWeapon {
     @Override
     public String status() {
         if (chargeCap == 100)
-            return Messages.format("%d%%", charge);
+            return Messages.format("%d%%", (int)charge);
         return null;
     }
 }
