+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/LoneJourney.java
@@ -120,7 +120,7 @@ public class LoneJourney extends MeleeWeapon {
         }
         //display as percent
         if (chargeCap == 100)
-            return Messages.format("%d%%", charge);
+            return Messages.format("%d%%", (int)charge);
 
 
         //otherwise, if there's no charge, return null.
