+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Suffering.java
@@ -224,7 +224,7 @@ public class Suffering extends MeleeWeapon {
     }
 
     @Override
-    public void SPCharge(int value) {
+    public void SPCharge(float value) {
         if(setbouns()) value *= 2;
         super.SPCharge(value);
     }
@@ -240,7 +240,7 @@ public class Suffering extends MeleeWeapon {
     @Override
     public String status() {
         if (chargeCap == 100)
-            return Messages.format("%d%%", charge);
+            return Messages.format("%d%%", (int)charge);
         return null;
     }
